finishing basic permissions features

+++ routes/urls.js
@@ -23,38 +23,54 @@ const deleteUrl = (shortURL) => {
 
 // LIST of URLS
 router.get('/urls', (req, res) => {
-  let templateVars = { 
-    user: findUserByID(req.session["user_id"]),
-    urls: findUrlsUser(user.id)
+  if (req.session["user_id"]) {
+    let templateVars = { 
+      user: findUserByID(req.session["user_id"]),
+      urls: findUrlsUser(user.id)
+    };
+    res.render("urls_index", templateVars);
+  } else {
+    res.status(401).send('Unauthorized, please <a href="/"> Login </a>');
   };
-  res.render("urls_index", templateVars);
 });
 
 // Display the form to add a new URL
 router.get('/urls/new', (req, res) => {
-  let templateVars = { 
-    user: findUserByID(req.session["user_id"]),
+  if (req.session["user_id"]) {
+    let templateVars = { 
+      user: findUserByID(req.session["user_id"]),
+    };
+    res.render("urls_new", templateVars);
+  } else {
+    res.status(401).send('Unauthorized, please <a href="/"> Login </a>');
   };
-  res.render("urls_new", templateVars);
 });
 
 // Add the new URL
 router.post('/urls',(req,res) => {
-  let shortURL = generateRandomString();
-  let longURL = req.body.longURL;
-  let user = req.session.user_id;
-  addNewUrl(shortURL,longURL, user);
-  res.redirect(`/urls/${shortURL}`);
+  if (req.session["user_id"]) {
+    let shortURL = generateRandomString();
+    let longURL = req.body.longURL;
+    let user = req.session.user_id;
+    addNewUrl(shortURL,longURL, user);
+    res.redirect(`/urls/${shortURL}`);
+  } else {
+    res.status(401).send('Unauthorized, please <a href="/"> Login </a>');
+  }
 });
 
 // Display a specific URL
 router.get('/urls/:shortURL', (req, res) => {
-  let templateVars = { 
-    user: findUserByID(req.session["user_id"]),
-    shortURL: req.params.shortURL, 
-    longURL: urlDatabase[req.params.shortURL]
+  if (req.session["user_id"]) {
+    let templateVars = { 
+      user: findUserByID(req.session["user_id"]),
+      shortURL: req.params.shortURL, 
+      longURL: urlDatabase[req.params.shortURL]
+    };
+    res.render("urls_show", templateVars);
+  } else {
+    res.status(401).send('Unauthorized, please <a href="/"> Login </a>');
   };
-  res.render("urls_show", templateVars);
 });
 
 // Redirect to the URL
@@ -75,19 +91,27 @@ router.get('/urls/:shortURL/update', (req, res) => {
 
 // Update the specific URL
 router.post('/urls/:shortURL', (req, res) => {
-  let user = req.session.user_id;
-  console.log(user);
-  let shortURL = req.params.shortURL;
-  let longURL = req.body.longURL;
-  updateUrl(shortURL, longURL, user);
-  res.redirect('/urls');
+  if (req.session["user_id"]) {
+    let user = req.session.user_id;
+    console.log(user);
+    let shortURL = req.params.shortURL;
+    let longURL = req.body.longURL;
+    updateUrl(shortURL, longURL, user);
+    res.redirect('/urls');
+  } else {
+    res.status(401).send('Unauthorized, please <a href="/"> Login </a>');
+  }
 });
 
 // Delete a specific URL
 router.post('/urls/:shortURL/delete', (req, res) => {
-  let shortURL = req.params.shortURL;
-  deleteUrl(shortURL);
-  res.redirect('/urls/');
+  if (req.session["user_id"]) {
+    let shortURL = req.params.shortURL;
+    deleteUrl(shortURL);
+    res.redirect('/urls/');
+  } else {
+    res.status(401).send('Unauthorized, please <a href="/"> Login </a>');
+  }
 });
 
 module.exports = router;
